BytesToString: Use int64

+++ includes/BytesToString/src/BytesToString/dllmain.cpp
@@ -8,7 +8,7 @@
 
 //#define VERBOSE
 
-static char* humanSize(int32_t bytes)
+static char* humanSize(int64_t bytes)
 {
     const char* suffix[] = { "B", "KB", "MB", "GB", "TB" };
     char length = sizeof(suffix) / sizeof(suffix[0]);
@@ -26,7 +26,7 @@ static char* humanSize(int32_t bytes)
     return output;
 }
 
-extern "C" __declspec(dllexport) const char* __cdecl BytesToString(int32_t size)
+extern "C" __declspec(dllexport) const char* __cdecl BytesToString(int64_t size)
 {
     #ifdef VERBOSE
     std::cout << humanSize(size) << std::endl;
